Refactor: Make more optional fields for objects (#231)
In a separate branch I am adding unit tests but I noticed the code to
mock objects is pretty big because of the amount of fields. Fortunately
we can set most of those fields to be optional and save us needing to
declare so many.

Tested by `npm run build` and also using the website to make sure it
still works fine.