Bugfix: Fix bugs in object details and searchbar
This fixes 2 big bugs.

When you open a details modal it used to automatically update the object type -- but that was confusing because when you closed it the object type had changed. So this removes that listener and instead I try to make the update in the correct buttons.

Secondarily, when you search and click a suggestion it open failed. I think the problem was because the searchbar is now in a form and Chrome thinks that the first button is a form submit that should refresh the page. I changed the button type and that no longer happens