Bugfix: Fix bugs in object details and searchbar (#235)
This fixes 2 big bugs.

When you open a details modal it used to automatically update the object
type -- but that was confusing because when you closed it the object
type had changed. So this removes that listener and instead I try to
make the update in the correct buttons.

Secondarily, when you search and click a suggestion it open failed. I
think the problem was because the searchbar is now in a form and Chrome
thinks that the first button is a form submit that should refresh the
page. I changed the button type and that no longer happens

I updated the search bar style a bit too since it was inconsistent.

|Before|After|
|--|--|
|<img width="367" height="56" alt="Screenshot 2025-10-01 at 10 32 29"
src="https://github.com/user-attachments/assets/f2050c47-e69f-4f5a-a5d8-a61f8a3db475"
/>|<img width="369" height="60" alt="Screenshot 2025-10-01 at 10 29 43"
src="https://github.com/user-attachments/assets/7cae1393-9f9a-4bf0-973a-57ce5cd1c308"
/>