Locales: Clean up TSV (#245)
The core mission of this change was to clean up the locale.tsv file as
we've improved census data coverage. While I did this I worked on a few
things to make it easier to audit the locales in our table.

TSV changes
* Sorted by population (then locale code)
* Added more locales that made sense now that we have census data,
mostly diaspora populations but also macrolanguages and dialects that we
missing.
* Updated the population numbers in here when the census values were
better. Ultimately we still want to load census data, but this is a good
start.
* Updated the Population Source enum, removing the numbers (that make
sense for sorting when it was a spreadsheet). Changed the names a bit
too.
* Updated some census tsvs to count for languages with multiple codes
better

There are some UI updates too
* Locales with null data show no population estimate instead of NaN
* Added a new sort key "PercentOfGlobalLanguageSpeakers" for locales
* Improved the copy button from the PotentialLocales table

Everything looks okay in the locale table view
<img width="876" height="812" alt="Screenshot 2025-10-08 at 09 21 06"
src="https://github.com/user-attachments/assets/e6fa1b69-e44e-4044-ba63-69f4a41b6aa3"
/>