Testing: Add tests for sorting & object accessors (#253)
The original mission of this PR was to add tests for other sort types
and for the object accessors -- it ended up growing pretty big.

Key changes:
* Made `MockObjects.tsx` -- a good way to mock all kinds of objects to
test sorting and accessors. Based on languages in JRR Tolkien's
literature to avoid using real languages.
* Used those mocks to create 3 new test files: `sort.test.tsx`,
`getObjectPopulation.test.tsx`, and `getObjectMiscFields.text.tsx` --
this will have merge conflicts with @abdulkalam556 's current PR so I
plan to rebase this change on that to make any necessary updates.

Other changes on the way:
* Moved `getTerritoryBiggestLocale` and
`getObjectMostImportantLanguageName` from `getObjectPopulation.tsx` to
`getObjectMiscFields.tsx` for better organization -- it probably won't
be the final location but at least it focuses getObjectPopulation
better.
* Added `populationFromUN` field to `TerritoryData` to distinguish
original UN data from recomputed values
* Split `getObjectDate` and `getObjectDateAsNumber` since 1 is used for
displaying and 1 is used for sorting (before it sorted the year 2000 as
2000 but as a dateNumber it should be 946684800000 (that many
millseconds after 1971).
* Compute regional locale literacy as well in
`computeLocaleWritingPopulation`
* Discovered while writing the tests, I fixed various accessors to fill
out values that weren't provided before.
* `getCountOfLanguages` for territories doesn't double-count languages
with multiple locales in a territory
  * `getObjectPopulation` will return undefined for variant tags
* `getObjectPopulationAttested` returns values for territories &
censuses
* `getObjectPopulationPercentInBiggestDescendentLanguage` returns
percents now

---------

Co-authored-by: conradarcturus <1616639+conradarcturus@users.noreply.github.com>