Tables: Column selection page param (#257)
While developing I was getting annoyed that column selection was not
persisting. This PR makes the column selection persist for your session
(for your particular tab)

# Background 
I considered making column selection a URL parameter but it was a bit
difficult to scale. The most first successful change was to make column
selection a binary number (each digit corresponds to if a column is off
or on) & put the binary number in the URL (converted to decimal, but we
could have done alphanumeric). This has the benefits of being stored in
the URL and part of the page history.

Unfortunately I realized it controlled columns in all tables, so if you
selected column 1, 2, and 5 in the LanguageTable and opened a
TerritoryTable, then it stayed with 1,2,5 selected. I considered making
the page parameters reset if you change ObjectType -- but the Table view
can be opened for multiple table types, even multiple table for the same
object type. Scaling this sounded impossible inside a URL.

Another thing I considered was encoding the columnVisibility array using
encoding functions -- but I similarily didn't think it would scale.

Thereby, I gave up on using the URL parameters and, instead, am using
sessionStorage. localStorage and sessionStorage are almost the same
thing -- but localStorage persists across tabs, sessionStorage is only
in the same tab.

# Changes

* Moved column visibility functions to `useColumnVisibility` hook.
* Added `useStoredParams` hook
* Changed the storage parameter from `useState` to `useStoredParams`
* Added a button to reset visibility.

# Testing

I added some vitest tests and I also used the website:
* Go to Table view http://localhost:5173/lang-nav/data?view=Table
* Open column selection, click on/off some columns, including common
ones like Name
* Changed object type to see the unique ones not change
* Refresh page, verified its persistent
* Open new tabs, verified changes don't affect those tabs

---------

Co-authored-by: conradarcturus <1616639+conradarcturus@users.noreply.github.com>