Table column state in URL (#307)
By golly, I did it. I was able to figure a compact way to encourage the
currently visible columns in the URL. I convert the visible column flags
to a binary (1 = visible, 0 = not visible), so 5 = 101 = visible/not
visible/visible.

Now you can go back/forward with column selection and share links :D

---------

Co-authored-by: conradarcturus <1616639+conradarcturus@users.noreply.github.com>