Refactor: Simplify language source specific fields (#312)
The LanguageData fields are simultaneously too nested and not enough :p
We have some ISO and CLDR fields on the base and we keep on repeating
"...sourceSpecific..." This changes them to balance them out a little
bit.

While I was here I also revised the references to the ISO vitality --
since it really is ISO status, not a fine-grained vitality measure.