Refactor: collect transforms (sorting, coloring, filtering) (#316)
I was writing up a task about adding a "scaleBy" parameter and I was
able to write "put it in the sorting folder like I did for coloring" but
I realized they should really be in separate folders -- even if there is
common approaches. So I changed them to have a different common folder
`@features/transforms` with 3 subfolders to start, `sorting`,
`coloring`, and `filtering`.

---------

Co-authored-by: conradarcturus <1616639+conradarcturus@users.noreply.github.com>