Language: Update population algorithm, especially wrt descendants
Now that we have a lot of new data coming in and new algorithms computing population -- unfortunately it just no longer makes sense to cache the population of languages from their descendants. With locale populations recalculating based on a myriad of factors, it is better to just compute it when updating the population source.