Sorting: Review getApplicableSortBy values (#332)
While I was talking with @cchandram about ScaleBy values and
re-organizing how the "getApplicable..." methods work, I wanted to try
poking holes in `getSortBysApplicableToObjectType`.

1) We can set a few common sortBys for all objects: Name, Code &
Population. Unfortunately not object types support endonyms since we
don't have data for endonyms for Variant Tags or Censuses.
2) Added a test for the function -- most interestingly revealing the
objects we had truthy values for that weren't listed in the SortBy.
3) Added some more territory names from my previous change adding
territory endonyms -- oof I wanted to add this to that change but I
committed them to the wrong branch.