Censuses: Update Russia data (#336)
I was verifying census data for Russia for a project so I just reviewed
the languages. I realized I had missed a table in the data I had
downloaded so this adds a new one and it splits the files up by the
data.un.org one and the official document.